refactor: create capability permission form
