+++ apps/torus-portal/src/app/(pages)/signals/create-signal/_components/create-signal-fields/create-signal-slider-field.tsx
